refactor(components/layout/grid): add import file extensions
