+++ components/layout/grid/demo/ArticleAlignContent.js
@@ -12,9 +12,9 @@ import {
 import LayoutGrid, {
   LayoutGridAlignContent,
   LayoutGridItem
-} from 'components/layout/grid/src'
-import DemoWrapper from './demoWrapper'
-import DemoBox from './demoBox'
+} from 'components/layout/grid/src/index.js'
+import DemoWrapper from './demoWrapper.js'
+import DemoBox from './demoBox.js'
 
 const ArticleAlignContent = ({classname}) => {
   const [alignContentState, setAlignContentState] = useState()
