+++ components/layout/grid/demo/ArticleCombine.js
@@ -10,9 +10,9 @@ import {
   RadioButton
 } from '@s-ui/documentation-library'
 
-import LayoutGrid, {LayoutGridItem} from 'components/layout/grid/src'
-import DemoWrapper from './demoWrapper'
-import DemoBox from './demoBox'
+import LayoutGrid, {LayoutGridItem} from 'components/layout/grid/src/index.js'
+import DemoWrapper from './demoWrapper.js'
+import DemoBox from './demoBox.js'
 
 const Sites = {
   BarSite: ({left, right}) => {
