+++ components/layout/grid/demo/ArticleGutter.js
@@ -1,9 +1,9 @@
 import PropTypes from 'prop-types'
 import {H2, Paragraph, Article, Code, Text} from '@s-ui/documentation-library'
 
-import LayoutGrid, {LayoutGridItem} from 'components/layout/grid/src'
-import DemoWrapper from './demoWrapper'
-import DemoBox from './demoBox'
+import LayoutGrid, {LayoutGridItem} from 'components/layout/grid/src/index.js'
+import DemoWrapper from './demoWrapper.js'
+import DemoBox from './demoBox.js'
 
 const ArticleGutter = ({classname}) => {
   return (
