+++ components/layout/grid/demo/ArticleResponsive.js
@@ -11,9 +11,9 @@ import {
 } from '@s-ui/documentation-library'
 import {useWindowSize} from 'react-use'
 
-import LayoutGrid, {LayoutGridItem} from 'components/layout/grid/src'
-import DemoWrapper from './demoWrapper'
-import DemoBox from './demoBox'
+import LayoutGrid, {LayoutGridItem} from 'components/layout/grid/src/index.js'
+import DemoWrapper from './demoWrapper.js'
+import DemoBox from './demoBox.js'
 
 const breakpoints = {
   xxs: 0,
