+++ components/layout/grid/test/index.test.js
@@ -7,10 +7,10 @@ import ReactDOM from 'react-dom'
 import chai, {expect} from 'chai'
 import chaiDOM from 'chai-dom'
 
-import * as pkg from '../src'
+import * as pkg from '../src/index.js'
 import json from '../package.json'
 
-import {getColSpanClassNamesTransform} from 'components/layout/grid/src/gridItem'
+import {getColSpanClassNamesTransform} from 'components/layout/grid/src/gridItem/settings.js'
 
 chai.use(chaiDOM)
